requirements.txt
Flask==3.1.2
requests==2.31.0
python-dotenv==1.0.1
boto3==1.30.0
redis==5.3.0